Improved file organisation with introduction of caseProgression directory
