+++ src/main/services/features/dashboard/claimSummary/latestUpdate/caseProgression/latestUpdateContentBuilderCaseProgression.ts
@@ -1,0 +1,11 @@
+import {Claim} from 'models/claim';
+import {ClaimSummarySection} from 'form/models/claimSummarySection';
+import {
+  getEvidenceUpload,
+} from 'services/features/dashboard/claimSummary/latestUpdate/latestUpdateContent/evidenceUploadContent';
+
+export const buildEvidenceUploadSection = (claim: Claim): ClaimSummarySection[] => {
+  const sectionContent = [];
+  sectionContent.push(getEvidenceUpload(claim));
+  return sectionContent.flat();
+};
