+++ src/main/services/features/dashboard/claimSummary/latestUpdateService.ts
@@ -1,6 +1,8 @@
 import {ClaimSummaryContent} from 'form/models/claimSummarySection';
 import {Claim} from 'models/claim';
-import {buildEvidenceUploadSection, buildResponseToClaimSection} from './latestUpdate/latestUpdateContentBuilder';
+import {
+  buildResponseToClaimSection,
+} from './latestUpdate/latestUpdateContentBuilder';
 
 export const getLatestUpdateContent = (claimId: string, claim: Claim): ClaimSummaryContent[] => {
   const responseToClaimSection = buildResponseToClaimSection(claim, claimId);
@@ -13,13 +15,4 @@ export const getLatestUpdateContent = (claimId: string, claim: Claim): ClaimSumm
       hasDivider: index < filteredLatestUpdateContent.length - 1,
     });
   });
-};
-
-export const getEvidenceUploadLatestUpdateContent = (claimId: string, claim: Claim): ClaimSummaryContent[] => {
-  const evidenceUploadSection = buildEvidenceUploadSection(claim);
-  const latestUpdateContent = [evidenceUploadSection];
-  return latestUpdateContent.map((sectionContent, index) => ({
-    contentSections: sectionContent,
-    hasDivider: index < evidenceUploadSection.length - 1,
-  }));
 };
